Split account server into partials
